LayoutTests/fast/forms/textarea-newline.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
These tests check if leading line feeds characters in textarea as default values are parsed correctly.

<br>
<textarea id="no-line-feed">Madoka</textarea>

<textarea id="one-line-feed">
Sayaka</textarea>

<textarea id="two-line-feeds">

Mami</textarea>

<textarea id="one-line-feed-escaped-char-and-one-line-feed">&#x0a;
Kyoko</textarea>

<textarea id="two-line-feed-escaped-chars">&#x0a;&#x0a;Homura</textarea>

<p id="description"></p>
<div id="console"></div>

<script>
shouldBe('document.getElementById("no-line-feed").value', '"Madoka"');

shouldBe('document.getElementById("one-line-feed").value', '"Sayaka"');

shouldBe('document.getElementById("two-line-feeds").value', '"\\nMami"');

shouldBe('document.getElementById("one-line-feed-escaped-char-and-one-line-feed").value', '"\\nKyoko"');

shouldBe('document.getElementById("two-line-feed-escaped-chars").value', '"\\nHomura"');
</script>

<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
